Moved provision code to its own file to clean up the client.rs.
